Fix stdint.h self-inclusion
On systems that provide a stdint.h, atom/src/stdint.h was supposed to include
the system's version.  However, it would recursively include itself, which is
a no-op since it's protected by a '#pragma once'.
The end result was that the system's stdint.h was not included.

Rename the local stdint.h "forwarder" to atomstdint.h and change the one file
that includes it.